backend/requirements/base.txt
Django==5.0.2
drf-spectacular==0.27.2
djangorestframework==3.15.1
django-filter==24.2
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
django-environ==0.12.0
django-minio-storage==0.5.7
djangorestframework_simplejwt==5.5.0